+++ src/SignalR/clients/cpp/test/signalrclienttests/websocket_transport_tests.cpp
@@ -11,81 +11,6 @@
 
 using namespace signalr;
 
-template <typename T>
-class manual_reset_event
-{
-public:
-    void set(T value)
-    {
-        m_promise.set_value(value);
-    }
-
-    void set_exception(std::exception exception)
-    {
-        m_promise.set_exception(std::make_exception_ptr(exception));
-    }
-
-    void set_exception(std::exception_ptr exception)
-    {
-        m_promise.set_exception(exception);
-    }
-
-    T get()
-    {
-        // TODO: timeout
-        try
-        {
-            auto ret = m_promise.get_future().get();
-            m_promise = std::promise<T>();
-            return ret;
-        }
-        catch (...)
-        {
-            m_promise = std::promise<T>();
-            std::rethrow_exception(std::current_exception());
-        }
-    }
-private:
-    std::promise<T> m_promise;
-};
-
-template <>
-class manual_reset_event<void>
-{
-public:
-    void set()
-    {
-        m_promise.set_value();
-    }
-
-    void set_exception(std::exception exception)
-    {
-        m_promise.set_exception(std::make_exception_ptr(exception));
-    }
-
-    void set_exception(std::exception_ptr exception)
-    {
-        m_promise.set_exception(exception);
-    }
-
-    void get()
-    {
-        try
-        {
-            m_promise.get_future().get();
-        }
-        catch (...)
-        {
-            m_promise = std::promise<void>();
-            std::rethrow_exception(std::current_exception());
-        }
-
-        m_promise = std::promise<void>();
-    }
-private:
-    std::promise<void> m_promise;
-};
-
 TEST(websocket_transport_connect, connect_connects_and_starts_receive_loop)
 {
     auto connect_called = false;
